Restructure file set up for organization
