+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import Login from './Login';
+import App from './App/App';
+import Login from './Login/Login';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
